apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.0.3
langchain==0.0.350
chromadb==0.4.18
openai==1.3.0
groq==0.4.0
pandas==2.1.3
openpyxl==3.1.2
python-magic==0.4.27
gTTS==2.4.0
python-jose==3.3.0
authlib==1.2.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
pytest-xdist==3.5.0
black==23.11.0
isort==5.12.0
mypy==1.7.0
ruff==0.1.6
